UPDATE 1.9.5 Preventing Session Hijacking by verifying the client's ip address

+++ Models/User.php
@@ -164,6 +164,7 @@ class User extends Password
                 $_SESSION['Account'] = $account;
                 $this->Mail->send($this->getMailAddress(), "Verification Needed!", "Your one-time password is {$this->getOtp()}.  Please use this password to complete the log in process on http://{$_SERVER['HTTP_HOST']}/Login/Verification/{$this->getUsername()}");
                 $data = array(
+                    "Client" => $_SESSION['Client'],
                     "User" => $_SESSION['User'],
                     "Account" => $_SESSION['Account']
                 );
@@ -199,16 +200,28 @@ class User extends Password
      */
     public function logOut()
     {
-        unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Session/Users/{$_SESSION['User']['username']}.json");
-        if (file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Profiles/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json") && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Match Histories/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json") && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Champion Masteries/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json")) {
-            unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Profiles/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
-            unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Match Histories/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
-            unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Champion Masteries/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
-            $response = array(
-                "status" => 0,
-                "url" => "http://{$_SERVER['HTTP_HOST']}",
-                "message" => "You have been successfully logged out!"
-            );
+        if (isset($_SESSION)) {
+            if (isset($_SESSION['Account']['LeagueOfLegends']) && isset($_SESSION['User']) && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Profiles/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json") && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Match Histories/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json") && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Champion Masteries/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json") && file_exists("{$_SERVER['DOCUMENT_ROOT']}/Cache/Session/Users/{$_SESSION['User']['username']}.json")) {
+                unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Profiles/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
+                unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Match Histories/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
+                unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Riot Games/Users/Champion Masteries/{$_SESSION['Account']['LeagueOfLegends']['playerUniversallyUniqueIdentifier']}.json");
+                unlink("{$_SERVER['DOCUMENT_ROOT']}/Cache/Session/Users/{$_SESSION['User']['username']}.json");
+                session_unset();
+                session_destroy();
+                $response = array(
+                    "status" => 0,
+                    "url" => "http://{$_SERVER['HTTP_HOST']}",
+                    "message" => "You have been successfully logged out!"
+                );
+            } else {
+                session_unset();
+                session_destroy();
+                $response = array(
+                    "status" => 13,
+                    "url" => "http://{$_SERVER['HTTP_HOST']}",
+                    "message" => "You have been successfully logged out but the cache has not been cleared on the application server!"
+                );
+            }
         } else {
             $response = array(
                 "status" => 13,
@@ -216,7 +229,6 @@ class User extends Password
                 "message" => "You have been successfully logged out but the cache has not been cleared on the application server!"
             );
         }
-        unset($_SESSION);
         header('Content-Type: application/json', true, 200);
         echo json_encode($response);
     }
